Created like functunality inside the document (upadating note)
