+++ src/components/note-editor/ActionBar.tsx
@@ -12,7 +12,7 @@ import { Button } from '@/components/ui/Button';
 import { DialogQuizz } from '@/components/note-editor/DialogQuizz';
 import { DialogGenerateContent } from './DialogGenerateContent';
 import { Spinner } from '@/components/ui/Spinner';
-import { LikeComponent } from '@/components/core/LikeComponent';
+import { LikeComponent } from '@/components/ui/LikeComponent';
 import { DialogImageOcr } from '@/components/note-editor/DialogImageOcr';
 
 // Store
@@ -20,15 +20,19 @@ import { useToastStore } from '@/store/useToastStore';
 import { DialogAskAI } from '@/components/note-editor/DialogAskAI';
 
 export const ActionBar: React.FC<{
+  noteId: string;
+  isLiked: boolean;
+  likeCount: number;
   isEditing: boolean;
   setIsEditing: React.Dispatch<React.SetStateAction<boolean>>;
   saveDocument: () => void;
   editor: EditorType;
   completionLoading: boolean;
-  noteId: string;
   allowEditing: boolean;
 }> = ({
   noteId,
+  isLiked,
+  likeCount,
   isEditing,
   setIsEditing,
   editor,
@@ -37,38 +41,18 @@ export const ActionBar: React.FC<{
   allowEditing,
 }) => {
   const user = useSession();
-
   const toast = useToastStore((state) => state.setToast);
-
-  const likeAction = async () => {
-    try {
-      await fetch('http://localhost:3000/api/core/home/notes/like', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          noteId,
-          userId: user.data?.user.id,
-          // exists: liked,
-        }), //image
-      });
-    } catch (error) {
-      console.error(error);
-    }
-  };
 
   return (
     <div className="flex items-center justify-between gap-4 overflow-scroll py-2">
       {!isEditing ? (
         <>
           <LikeComponent
-            hasBeenLiked={false}
-            numberOfLikes={330}
+            noteId={noteId}
+            userId={user.data?.user.id!} // Karlo: Create this to not be only visible to signed up users
+            numberOfLikes={likeCount}
+            hasBeenLiked={isLiked}
             size="lg"
-            action={() => {
-              console.log('Liked');
-            }}
           />
           {allowEditing && (
             <Button
